feat: improve chat flow, setup using ConversationChain and ConversationalRetrievalChain
ConversationalRetrievalChain allows user to chat with the doc wherein
additional context is provided via the vectorstore, when user doesn't
want document specific answers, ConversationChain is user.

Common memory is used to ensure sharing of the conversation history
among the two chains. A custom func to get chat history is implemented
for ConversationalRetrievalChain as it requires a list of BaseMessage objects
but ConversationChain directly gets string version